Finished. Needs polished with group
Everything is finished but needs to be polished with group. Respawn points are in place and floor is ready to be coded. Need to decide how to win the game in the end room. I know we discussed about a hat as a option to win and have pop up text.