BM and Keys
Background Music has been added and The keys can now be collected and the counter updates with each one that the player gets.  The win and door text is just turned off currently because both are currently just stuck in the middlee of my screen.  Also changed the UI canvas to be 1920x1080